+++ packages/helic/test/Main.hs
@@ -6,6 +6,7 @@ import Helic.Test.InsertEventTest (test_insertEvent)
 import Helic.Test.ListTest (test_list)
 import Helic.Test.ListenTest (test_listen)
 import Helic.Test.LoadTest (test_load)
+import Helic.Test.StreamTest (test_stream)
 import Polysemy.Test (unitTest)
 import Test.Tasty (TestTree, defaultMain, testGroup)
 
@@ -17,7 +18,8 @@ tests =
     unitTest "listen for events, filter duplicates from network feedback" test_listen,
     unitTest "print the history" test_list,
     unitTest "load an old event to the clipboard" test_load,
-    unitTest "restart the gtk main loop when requested after failure" test_gtkMain
+    unitTest "restart the gtk main loop when requested after failure" test_gtkMain,
+    unitTest "stream events over http" test_stream
   ]
 
 main :: IO ()